Add tremendously dubious text rendering system
This system works by regex parsing a string for all occurrences of tags
that represent bold `*`, italics `/`, or underline `_` and creating a
tree-like structure that holds the string and formatting data.

A basic parse to tree would be: This is *bold* text
  - {v: "This is "}
  - {t: "*", v: "bold"}
  - {v: " text"}

A more complicated one would be: This is /italicized and *bold*/!
  - {v: "This is "}
  - {t: '/', c: [{v: "italicized and "}, {t: '*', v: "bold"}]}
  - {v: '!'}

As such, any contained elements render with the entire styling
tree that it is within. This part is, obviously, done during the render
phase by passing the parent's styling state to the child.

p.s., it doesn't work exactly right in regards to word spacings -- some
goofy off-by-one error of some sort...