Auto-resize canvas to fit all elements
This allows all elements to be shown on screen. Still to-do is to use
rules similar to the Image element, such that the entire diagram shows
but can be scaled down by providing explicit sizing rules to the
uxf-canvas. It would also be good, I suspect, to add some sort of
overflow rules in additional, as one may want to have a large UXF
element embedded in a page that can be scrolled within.