Fix font-family parsing, add horrid text height guessing code, change text rendering, add broken zoom support
Text rendering now occurs with textBaseline="top", as this makes rendering easier.

For whatever reason, HTML5 canvas doesn't provide any way to get the height of text, so we're using some hidden dummyText to get the height. However, if you want to be amused at how weird Chrome is, open uxf-canvas.html, click on the body, reload, see the underlined text work fine, don't click on the body, reload, see the underlined text not work fine. Chrome, why are you influencing the rendering of the page of the NEXT reload based on user input on THIS page???(x9000)